BUG#32377134 SEGMENTATION FAULT WHEN ADDING EMPTY ARRAY TO THE COLLECTION
MySQL Shell crashes with segmentation fault after adding empty array to
the collection: `db.collection.add([])`.

Segmentation fault was caused by resultset nullptr dereference, and
resultset was nullptr, because calling db.collection.add with an empty
array as arugment do not send anything to the MySQL Server, therefore we
don't receive anything from server and resultset is not set.

Now, we skip dumping resultset information if resultset object is not
available.

+++ modules/mod_mysql_resultset.cc
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014, 2020, Oracle and/or its affiliates.
+ * Copyright (c) 2014, 2021, Oracle and/or its affiliates.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License, version 2.0,
@@ -562,7 +562,7 @@ shcore::Value ClassicResult::get_member(const std::string &prop) const {
         mysqlshdk::utils::format_seconds(_result->get_execution_time()));
 
   if (prop == "autoIncrementValue")
-    return shcore::Value((int)_result->get_auto_increment_value());
+    return shcore::Value(_result->get_auto_increment_value());
 
   if (prop == "info") return shcore::Value(_result->get_info());
